13/6

+++ example-app/app/Http/Controllers/MahasiswaController.php
@@ -14,7 +14,13 @@ class MahasiswaController extends Controller
      */
     public function index()
     {
-        $mahasiswa = Mahasiswa::all(); // select * from mahasiswa 
+        if(auth()->user()->role == 'D'){
+            $mahasiswa = Mahasiswa::where('user_id',auth()->user()->id)->get();
+            // select * from mahasiswa where user_id =1
+        } else {
+            $mahasiswa = Mahasiswa::all(); // select * from mahasiswa 
+        }
+        //$mahasiswa = Mahasiswa::all(); // select * from mahasiswa 
         return view('mahasiswa.index')
                 ->with('mahasiswa', $mahasiswa);
     }
@@ -90,6 +96,10 @@ class MahasiswaController extends Controller
      */
     public function update(Request $request, Mahasiswa $mahasiswa)
     {
+
+        if (auth()->user()->cannot('delete', $mahasiswa)) {
+            abort(403);
+        }
         //dd($request);
         if($request->url_foto){ // jika ada file yangdilampirkan 
             $val = $request ->validate([
@@ -138,6 +148,9 @@ class MahasiswaController extends Controller
      */
     public function destroy(Mahasiswa $mahasiswa)
     {
+        if (auth()->user()->cannot('delete', $mahasiswa)) {
+            abort(403);
+        }
         //dd($mahasiswa);
         File:: delete('foto/'. $mahasiswa['url_foto']);
         $mahasiswa->delete(); // untuk menghapus data mahasiswa
